move migration code before agree to privacy policies
